Update install-rcf-software-9336c-cluster.adoc
Removed the extra erroneous lines 8-13 that were confusing many customers and installers. It was showing the write erase information that should only be done when upgrading the RCF (different procedure), or if the RCF needs to be wiped and redone (not specified in this document). There is no mention that these included steps were for this. In addition, the step to copy the new RCF after the erase config is not included, so it would basically be wiping the RCF only. If the intent was to include this in the case where RCF needs to be wiped during install (for example, wrong RCF was used), then it should be a separate procedure added below the standard procedures. Currently, it reads as though it is a required step when initially installing and RCF for the first time.